Fix the last x64 crash: int VM_ArgPtr prototype in cl_ui.c truncated 64-bit pointers
cl_ui.c carried a local forward declaration `void *VM_ArgPtr( int intValue );`
that shadowed qcommon.h's `intptr_t` prototype. Inside the UI syscall handler,
VMA(x) = VM_ArgPtr(args[x]) therefore truncated every pointer argument to 32
bits at the call site. It went unnoticed because the UI module's globals load
below 4 GB (truncation is a no-op there), but the FIRST pointer to a high
address -- &arenasFile, a stack vmCvar_t in UI_LoadArenas during UI_INIT --
truncated to garbage and crashed Cvar_Register. cl_cgame.c / sv_game.c were
unaffected (they use qcommon.h's correct prototype), which is why only the UI
crashed.

With this, the x64 + Vulkan build runs the full game: engine, native game DLLs
(cgame/qagame/ui), the Vulkan renderer, q3_ui menus and sound all come up and
render (verified by an in-engine screenshot of the menu at 1920x1080).

Also removes the temporary VM_ArgPtr/cvar-register debug instrumentation. The
opt-in Com_BootLog tracer (gated by the Q3_BOOTLOG env var) is kept -- it is
inert in normal play and was invaluable for pinning these down.

+++ code/client/cl_ui.c
@@ -757,7 +757,7 @@ static int FloatAsInt( float f ) {
 	return temp;
 }
 
-void *VM_ArgPtr( int intValue );
+void *VM_ArgPtr( intptr_t intValue );	// must match qcommon.h; an int here truncated 64-bit pointers
 #define	VMA(x) VM_ArgPtr(args[x])
 #define	VMF(x)	(*(float *)&args[x])
 
@@ -782,7 +782,7 @@ intptr_t CL_UISystemCalls( intptr_t *args ) {
 		return Sys_Milliseconds();
 
 	case UI_CVAR_REGISTER:
-		Cvar_Register( VMA(1), VMA(2), VMA(3), args[4] ); 
+		Cvar_Register( VMA(1), VMA(2), VMA(3), args[4] );
 		return 0;
 
 	case UI_CVAR_UPDATE:
